コントローラーのディレクトリを整理した

+++ routes/web.php
@@ -14,7 +14,7 @@
 Route::redirect('/', '/auth/login');
 
 /* Top Page */
-Route::get('/top', 'HomeController@index')->name('top_page');
+Route::get('/top', 'Admin\HomeController@index')->name('top_page');
 
 /* Auth */
 Route::prefix('auth')->group(function () {
@@ -25,26 +25,26 @@ Route::prefix('auth')->group(function () {
 
 /* Admin Users */
 Route::prefix('admin_users')->group(function() {
-    Route::get('/','AdminUserController@index')->name('admin_users.index');
-    Route::get('/create','AdminUserController@create')->name('admin_users.create');
-    Route::post('/','AdminUserController@store')->name('admin_users.store');
-    Route::get('/{adminUser}/edit','AdminUserController@edit')->where('user', '[0-9]+')->name('admin_users.edit');
-    Route::put('/{adminUser}','AdminUserController@update')->where('user', '[0-9]+')->name('admin_users.update');
-    Route::delete('/{adminUser}','AdminUserController@destroy')->where('user', '[0-9]+')->name('admin_users.destroy');
+    Route::get('/','Admin\AdminUserController@index')->name('admin_users.index');
+    Route::get('/create','Admin\AdminUserController@create')->name('admin_users.create');
+    Route::post('/','Admin\AdminUserController@store')->name('admin_users.store');
+    Route::get('/{adminUser}/edit','Admin\AdminUserController@edit')->where('user', '[0-9]+')->name('admin_users.edit');
+    Route::put('/{adminUser}','Admin\AdminUserController@update')->where('user', '[0-9]+')->name('admin_users.update');
+    Route::delete('/{adminUser}','Admin\AdminUserController@destroy')->where('user', '[0-9]+')->name('admin_users.destroy');
 });
 
 /* Drugs */
 Route::prefix('drugs')->group(function() {
-   Route::get('/', 'DrugController@index')->name('drugs.index');
-   Route::get('/create', 'DrugController@create')->name('drugs.create');
-   Route::post('/', 'DrugController@store')->name('drugs.store');
-   Route::get('/edit/{drug}', 'DrugController@edit')->name('drugs.edit');
-   Route::post('/update/{drug}', 'DrugController@update')->name('drugs.update');
-   Route::post('/{drug}', 'DrugController@delete')->name('drugs.delete');
+   Route::get('/', 'Admin\DrugController@index')->name('drugs.index');
+   Route::get('/create', 'Admin\DrugController@create')->name('drugs.create');
+   Route::post('/', 'Admin\DrugController@store')->name('drugs.store');
+   Route::get('/edit/{drug}', 'Admin\DrugController@edit')->name('drugs.edit');
+   Route::post('/update/{drug}', 'Admin\DrugController@update')->name('drugs.update');
+   Route::post('/{drug}', 'Admin\DrugController@delete')->name('drugs.delete');
 });
 
 Route::prefix('medication_histories')->group(function(){
-   Route::get('/', 'MedicationHistoryController@index')->name('medication_histories.index');
-   Route::get('/edit/{medicationHistory}', 'MedicationHistoryController@edit')->name('medication_histories.edit');
-   Route::post('/update/{medicationHistory}', 'MedicationHistoryController@update')->name('medication_histories.update');
+   Route::get('/', 'Admin\MedicationHistoryController@index')->name('medication_histories.index');
+   Route::get('/edit/{medicationHistory}', 'Admin\MedicationHistoryController@edit')->name('medication_histories.edit');
+   Route::post('/update/{medicationHistory}', 'Admin\MedicationHistoryController@update')->name('medication_histories.update');
 });
